+++ core/services/worker/replica_test.go
@@ -1,4 +1,4 @@
-package cli
+package worker
 
 import (
 	. "github.com/onsi/ginkgo/v2"
@@ -30,11 +30,11 @@ var _ = Describe("Worker per-replica process keying", func() {
 
 	Describe("registrationBody", func() {
 		It("includes max_replicas_per_model and the auto-label", func() {
-			cmd := &WorkerCMD{
+			cfg := &Config{
 				Addr:                "worker.example.com:50051",
 				MaxReplicasPerModel: 4,
 			}
-			body := cmd.registrationBody()
+			body := cfg.registrationBody()
 
 			Expect(body).To(HaveKey("max_replicas_per_model"))
 			Expect(body["max_replicas_per_model"]).To(Equal(4))
@@ -45,8 +45,8 @@ var _ = Describe("Worker per-replica process keying", func() {
 		})
 
 		It("coerces zero/unset MaxReplicasPerModel to 1", func() {
-			cmd := &WorkerCMD{Addr: "worker.example.com:50051"}
-			body := cmd.registrationBody()
+			cfg := &Config{Addr: "worker.example.com:50051"}
+			body := cfg.registrationBody()
 			Expect(body["max_replicas_per_model"]).To(Equal(1),
 				"unset must default to single-replica behavior, not capacity 0")
 
@@ -55,12 +55,12 @@ var _ = Describe("Worker per-replica process keying", func() {
 		})
 
 		It("preserves user-provided labels alongside the auto-label", func() {
-			cmd := &WorkerCMD{
+			cfg := &Config{
 				Addr:                "worker.example.com:50051",
 				MaxReplicasPerModel: 2,
 				NodeLabels:          "tier=fast,gpu=a100",
 			}
-			body := cmd.registrationBody()
+			body := cfg.registrationBody()
 			labels := body["labels"].(map[string]string)
 			Expect(labels).To(HaveKeyWithValue("tier", "fast"))
 			Expect(labels).To(HaveKeyWithValue("gpu", "a100"))
